Add files via upload
French & Turkish translations for the latest version of Blocek. The problem is that some characters are garbled and aren't displayed properly ("è", "à", "À", and probably others in French, "ı", "ş", "ğ", "İ"  and probably others in Turkish). I've sent a message to Laaca, awaiting his response but uploading these in the meantime -- they'll probably be updated at some point.