Add files via upload
Alright, the garbled display is fixed amongst other issues thanks to Laaca's help. Now everything should be fine. Note that the language names are in UTF-8 in the files that are otherwise in IBM CP850 and 857, let's hope GitHub won't break that. The files entirely in UTF-8  are only there for convenience, the other ones need to be used in Blocek's translat directory.